+++ CHANGELOG.md
@@ -22,6 +22,8 @@
 🥷 *Internal*
 * Remove unused `_db` code
 
+* Change `RuntimeConfig._name` typing from `Optional[str]` to `str`
+
 📃 *Docs*
 
 * Update _Runtime Configuration_ guide and config-related docstrings to align them with the ``auto`` config.